// src/types.ts
export const TYPES = {
  Bot: Symbol("Bot"),
  Client: Symbol("Client"),
  Token: Symbol("Token"),
};

// Generated by https://quicktype.io

export interface QuranDataRespEn {
  verse: Verse;
}

export interface Verse {
  id: number;
  verse_number: number;
  verse_key: string;
  juz_number: number;
  hizb_number: number;
  rub_number: number;
  sajdah_type: null;
  sajdah_number: null;
  chapter_id: number;
  page_number: number;
  translations: Translation[];
}

export interface Translation {
  id: number;
  resource_id: number;
  text: string;
  resource_name: string;
}

export interface DataForEmbed {
  color?: number;
  title: string;
  description?: string;
  fields?: {
    name: string;
    value: string;
  }[];
  footer?: {
    text: string;
  };
}

export interface settingsData {
  Channel?: string;
  Time?: string;
  Timezone?: string;
  Prefix?: string;
}

// Generated by https://quicktype.io

export interface QuranDataRespAr {
  verses: VerseAr[];
}

export interface VerseAr {
  id: number;
  verse_key: string;
  text_uthmani: string;
}

export type Lang = "en" | "ar" | "mixed";